+++ Nodify/NodifyEditor.cs
@@ -428,14 +428,14 @@ namespace Nodify
 
         private void OnItemSelectCompleted()
         {
-            if (ItemsSelectCompletedCommand?.CanExecute(null) ?? false)
-                ItemsSelectCompletedCommand.Execute(null);
+            if (ItemsSelectCompletedCommand?.CanExecute(DataContext) ?? false)
+                ItemsSelectCompletedCommand.Execute(DataContext);
         }
 
         private void OnItemsSelectStarted()
         {
-            if (ItemsSelectStartedCommand?.CanExecute(null) ?? false)
-                ItemsSelectStartedCommand.Execute(null);
+            if (ItemsSelectStartedCommand?.CanExecute(DataContext) ?? false)
+                ItemsSelectStartedCommand.Execute(DataContext);
         }
 
         private Rect selectedArea;
@@ -902,6 +902,13 @@ namespace Nodify
                 onFinish?.Invoke();
             }
         }
+
+        /// <summary>
+        /// Moves the viewport center at the center of the specified area.
+        /// </summary>
+        /// <param name="area">The location in graph space coordinates.</param>
+        public new void BringIntoView(Rect area)
+            => BringIntoView(new Point(area.X + area.Width / 2, area.Y + area.Height / 2));
 
         /// <summary>
         /// Scales the viewport to fit the specified <paramref name="area"/> or all the <see cref="ItemContainer"/>s if that's possible.
@@ -1417,9 +1424,9 @@ namespace Nodify
                 ItemsHost.Panel?.InvalidateArrange();
             }
 
-            if (ItemsDragCompletedCommand?.CanExecute(null) ?? false)
+            if (ItemsDragCompletedCommand?.CanExecute(DataContext) ?? false)
             {
-                ItemsDragCompletedCommand.Execute(null);
+                ItemsDragCompletedCommand.Execute(DataContext);
             }
         }
 
@@ -1438,9 +1445,9 @@ namespace Nodify
 
             if (Selection.Count > 0)
             {
-                if (ItemsDragStartedCommand?.CanExecute(null) ?? false)
+                if (ItemsDragStartedCommand?.CanExecute(DataContext) ?? false)
                 {
-                    ItemsDragStartedCommand.Execute(null);
+                    ItemsDragStartedCommand.Execute(DataContext);
                 }
 
                 e.Handled = true;